+++ app/controllers/organizations/staff/external_form_upload_controller.rb
@@ -7,6 +7,17 @@ module Organizations
         authorize! :external_form_upload,
           context: {organization: Current.organization}
       end
+        
+      def create
+        authorize! :external_form_upload,
+          context: {organization: Current.organization}
+        files = params.require(:files)
+        files.each do |file|
+          next if not file.is_a?(ActionDispatch::Http::UploadedFile)
+          import_service = Organizations::Importers::GoogleCsvImportService.new(file)
+          import_service.call
+        end
+      end
     end
   end
 end
