+++ app/policies/organizations/external_form_upload_policy.rb
@@ -6,5 +6,9 @@ module Organizations
     def index?
       permission?(:manage_external_form_uploads)
     end
+
+    def create?
+      permission?(:manage_external_form_uploads)
+    end
   end
 end
